danet_sample/danet/requirement.txt
paddlepaddle
numpy
pandas
matplotlib
seaborn
Pillow
opencv-python